Refactor SVD to frontend
