+++ tensorly/backend/__init__.py
@@ -66,11 +66,11 @@ class BackendManager(types.ModuleType):
         "clip",
         "kr",
         "kron",
-        "partial_svd",
         "lstsq",
         "eps",
         "finfo",
         "solve",
+        "svd",
         "qr",
         "randn",
         "check_random_state",
@@ -82,8 +82,6 @@ class BackendManager(types.ModuleType):
         "norm",
         "to_numpy",
         "is_tensor",
-        "randomized_range_finder",
-        "randomized_svd",
         "argsort",
         "flip",
         "count_nonzero",
@@ -117,7 +115,6 @@ class BackendManager(types.ModuleType):
         "nan",
         "complex128",
         "complex64",
-        "SVD_FUNS",
         "index",
         "backend_name",
     ]
