+++ tensorly/backend/core.py
@@ -1052,418 +1052,13 @@ class Backend(object):
 
         return res * m
 
-    def svd_flip(self, U, V, u_based_decision=True):
-        """Sign correction to ensure deterministic output from SVD.
-        Adjusts the columns of u and the rows of v such that the loadings in the
-        columns in u that are largest in absolute value are always positive.
-        This function is borrowed from scikit-learn/utils/extmath.py
-        Parameters
-        ----------
-        U : ndarray
-            u and v are the output of `partial_svd`
-        V : ndarray
-            u and v are the output of `partial_svd`
-        u_based_decision : boolean, (default=True)
-            If True, use the columns of u as the basis for sign flipping.
-            Otherwise, use the rows of v. The choice of which variable to base the
-            decision on is generally algorithm dependent.
-        Returns
-        -------
-        u_adjusted, v_adjusted : arrays with the same dimensions as the input.
-        """
-        if u_based_decision:
-            # columns of U, rows of V
-            max_abs_cols = self.argmax(self.abs(U), axis=0)
-            signs = self.sign(
-                self.tensor(
-                    [U[i, j] for (i, j) in zip(max_abs_cols, range(self.shape(U)[1]))],
-                    **self.context(U),
-                )
-            )
-            U = U * signs
-            if self.shape(V)[0] > self.shape(U)[1]:
-                signs = self.concatenate(
-                    (signs, self.ones(self.shape(V)[0] - self.shape(U)[1]))
-                )
-            V = V * signs[: self.shape(V)[0]][:, None]
-        else:
-            # rows of V, columns of U
-            max_abs_rows = self.argmax(self.abs(V), axis=1)
-            signs = self.sign(
-                self.tensor(
-                    [V[i, j] for (i, j) in zip(range(self.shape(V)[0]), max_abs_rows)],
-                    **self.context(V),
-                )
-            )
-            V = V * signs[:, None]
-            if self.shape(U)[1] > self.shape(V)[0]:
-                signs = self.concatenate(
-                    (signs, self.ones(self.shape(U)[1] - self.shape(V)[0]))
-                )
-            U = U * signs[: self.shape(U)[1]]
-
-        return U, V
-
     def svd(self, matrix):
         raise NotImplementedError
 
     def eigh(self, matrix):
         raise NotImplementedError
 
-    def partial_svd(
-        self, matrix, n_eigenvecs=None, flip=False, random_state=None, **kwargs
-    ):
-        """Computes a fast partial SVD on `matrix`
-
-        If `n_eigenvecs` is specified, sparse eigendecomposition is used on
-        either matrix.dot(matrix.T) or matrix.T.dot(matrix).
-
-        Parameters
-        ----------
-        matrix : tensor
-            A 2D tensor.
-        n_eigenvecs : int, optional, default is None
-            If specified, number of eigen[vectors-values] to return.
-        flip : bool, default is False
-            If True, the SVD sign ambiguity is resolved by making the largest component
-            in the columns of U, positive.
-        random_state: {None, int, np.random.RandomState}
-            If specified, use it for sampling starting vector in a partial SVD(scipy.sparse.linalg.eigsh)
-        **kwargs : optional
-            kwargs are used to absorb the difference of parameters among the other SVD functions
-
-        Returns
-        -------
-        U : 2-D tensor, shape (matrix.shape[0], n_eigenvecs)
-            Contains the right singular vectors
-        S : 1-D tensor, shape (n_eigenvecs, )
-            Contains the singular values of `matrix`
-        V : 2-D tensor, shape (n_eigenvecs, matrix.shape[1])
-            Contains the left singular vectors
-        """
-        # Check that matrix is... a matrix!
-        if self.ndim(matrix) != 2:
-            raise ValueError(
-                "matrix be a matrix. matrix.ndim is %d != 2" % self.ndim(matrix)
-            )
-
-        ctx = self.context(matrix)
-        is_numpy = isinstance(matrix, np.ndarray)
-        if not is_numpy:
-            warnings.warn(
-                "In partial_svd: converting to NumPy."
-                " Check SVD_FUNS for available alternatives if you want to avoid this."
-            )
-
-        # Choose what to do depending on the params
-        dim_1, dim_2 = self.shape(matrix)
-        min_dim, max_dim = (dim_1, dim_2) if dim_1 < dim_2 else (dim_2, dim_1)
-        matrix = self.to_numpy(matrix)
-
-        if (n_eigenvecs is None) or (n_eigenvecs >= min_dim):
-            # Just perform trucated SVD
-            full_matrices = (n_eigenvecs is None) or (n_eigenvecs > min_dim)
-            # If n_eigenvecs == min_dim, we don't want full_matrices=True, it's super slow
-            U, S, V = scipy.linalg.svd(matrix, full_matrices=full_matrices)
-            U, S, V = U[:, :n_eigenvecs], S[:n_eigenvecs], V[:n_eigenvecs, :]
-        else:
-            # We can perform a partial SVD
-            rng = self.check_random_state(random_state)
-            # initilize with [-1, 1] as in ARPACK
-            v0 = rng.uniform(-1, 1, min_dim)
-
-            # First choose whether to use X * X.T or X.T *X
-            if dim_1 < dim_2:
-                S, U = scipy.sparse.linalg.eigsh(
-                    np.dot(matrix, matrix.T.conj()), k=n_eigenvecs, which="LM", v0=v0
-                )
-                S = np.sqrt(np.clip(S, 0, None))
-                S = np.clip(
-                    S, np.finfo(S.dtype).eps, None
-                )  # To avoid divide by zero warning on next line
-                V = np.dot(
-                    matrix.T.conj(),
-                    U * np.where(np.abs(S) <= np.finfo(S.dtype).eps, 0, 1 / S)[None, :],
-                )
-                U, S, V = U[:, ::-1], S[::-1], V[:, ::-1]
-                V, R = np.linalg.qr(V)
-                V = V * (
-                    2 * (np.diag(R) >= 0) - 1
-                )  # we can't use np.sign because np.sign(0) == 0
-            else:
-                S, V = scipy.sparse.linalg.eigsh(
-                    np.dot(matrix.T.conj(), matrix), k=n_eigenvecs, which="LM", v0=v0
-                )
-                S = np.sqrt(np.clip(S, 0, None))
-                S = np.clip(S, np.finfo(S.dtype).eps, None)
-                U = (
-                    np.dot(matrix, V)
-                    * np.where(np.abs(S) <= np.finfo(S.dtype).eps, 0, 1 / S)[None, :]
-                )
-                U, S, V = U[:, ::-1], S[::-1], V[:, ::-1]
-                U, R = np.linalg.qr(U)
-                U = U * (2 * (np.diag(R) >= 0) - 1)
-
-            # WARNING: here, V is still the transpose of what it should be
-            V = V.T.conj()
-
-        if flip:
-            U, V = self.svd_flip(U, V)
-
-        if not is_numpy:
-            U = self.tensor(U, **ctx)
-            S = self.tensor(S, **ctx)
-            V = self.tensor(V, **ctx)
-
-        return U, S, V
-
-    def truncated_svd(self, matrix, n_eigenvecs=None, **kwargs):
-        """Computes a truncated SVD on `matrix` using the backends's standard SVD
-
-        Parameters
-        ----------
-        matrix : 2D-array
-        n_eigenvecs : int, optional, default is None
-            if specified, number of eigen[vectors-values] to return
-
-        Returns
-        -------
-        U : 2D-array
-            of shape (matrix.shape[0], n_eigenvecs)
-            contains the right singular vectors
-        S : 1D-array
-            of shape (n_eigenvecs, )
-            contains the singular values of `matrix`
-        V : 2D-array
-            of shape (n_eigenvecs, matrix.shape[1])
-            contains the left singular vectors
-        """
-        # Check that matrix is... a matrix!
-        if self.ndim(matrix) != 2:
-            raise ValueError(
-                "matrix be a matrix. matrix.ndim is %d != 2" % self.ndim(matrix)
-            )
-
-        dim_1, dim_2 = self.shape(matrix)
-        min_dim, max_dim = min(dim_1, dim_2), max(dim_1, dim_2)
-
-        if n_eigenvecs is None:
-            n_eigenvecs = max_dim
-
-        if n_eigenvecs > max_dim:
-            warnings.warn(
-                "Trying to compute SVD with n_eigenvecs={0}, which "
-                "is larger than max(matrix.shape)={1}. Setting "
-                "n_eigenvecs to {1}".format(n_eigenvecs, max_dim)
-            )
-            n_eigenvecs = max_dim
-
-        full_matrices = n_eigenvecs > min_dim
-
-        U, S, V = self.svd(matrix, full_matrices=full_matrices)
-        U, S, V = U[:, :n_eigenvecs], S[:n_eigenvecs], V[:n_eigenvecs, :]
-        return U, S, V
-
-    def symeig_svd(self, matrix, n_eigenvecs=None, **kwargs):
-        """Computes a truncated SVD on `matrix` using symeig
-
-            Uses symeig on matrix.T.dot(matrix) or its transpose
-
-        Parameters
-        ----------
-        matrix : 2D-array
-        n_eigenvecs : int, optional, default is None
-            if specified, number of eigen[vectors-values] to return
-        **kwargs : optional
-            kwargs are used to absorb the difference of parameters among the other SVD functions
-
-        Returns
-        -------
-        U : 2D-array
-            of shape (matrix.shape[0], n_eigenvecs)
-            contains the right singular vectors
-        S : 1D-array
-            of shape (n_eigenvecs, )
-            contains the singular values of `matrix`
-        V : 2D-array
-            of shape (n_eigenvecs, matrix.shape[1])
-            contains the left singular vectors
-        """
-        # Check that matrix is... a matrix!
-        if self.ndim(matrix) != 2:
-            raise ValueError(
-                "matrix be a matrix. matrix.ndim is %d != 2" % self.ndim(matrix)
-            )
-
-        dim_1, dim_2 = self.shape(matrix)
-        min_dim, max_dim = min(dim_1, dim_2), max(dim_1, dim_2)
-
-        if n_eigenvecs is None:
-            n_eigenvecs = max_dim
-
-        if n_eigenvecs > max_dim:
-            warnings.warn(
-                "Trying to compute SVD with n_eigenvecs={0}, which "
-                "is larger than max(matrix.shape)={1}. Setting "
-                "n_eigenvecs to {1}".format(n_eigenvecs, max_dim)
-            )
-            n_eigenvecs = max_dim
-
-        if dim_1 > dim_2:
-            S, U = self.eigh(self.dot(matrix, self.transpose(matrix)))
-            S = self.sqrt(self.clip(S, self.eps(S.dtype)))
-            V = self.dot(self.transpose(matrix), U / self.reshape(S, (1, -1)))
-        else:
-            S, V = self.eigh(self.dot(self.transpose(matrix), matrix))
-            S = self.sqrt(self.clip(S, self.eps(S.dtype)))
-            U = self.dot(matrix, V) / self.reshape(S, (1, -1))
-
-        U, S, V = (
-            self.flip(U, axis=1),
-            self.flip(S),
-            self.flip(self.transpose(V), axis=0),
-        )
-        return (
-            U[:, : min(dim_1, n_eigenvecs)],
-            S[: min(dim_1, dim_2, n_eigenvecs)],
-            V[: min(dim_2, n_eigenvecs), :],
-        )
-
-    def randomized_svd(
-        self,
-        matrix,
-        n_eigenvecs=None,
-        n_oversamples=5,
-        n_iter=2,
-        random_state=None,
-        **kwargs,
-    ):
-        """Computes a truncated randomized SVD.
-
-        If `n_eigenvecs` is specified, sparse eigendecomposition is used on
-        either matrix.dot(matrix.T) or matrix.T.dot(matrix).
-
-        Parameters
-        ----------
-        matrix : tensor
-            A 2D tensor.
-        n_eigenvecs : int, optional, default is None
-            If specified, number of eigen[vectors-values] to return.
-        n_oversamples: int, optional, default = 5
-            rank overestimation value for finiding subspace with better allignment
-        n_iter: int, optional, default = 2
-            number of power iterations for the `randomized_range_finder` subroutine
-        random_state: {None, int, np.random.RandomState}
-        **kwargs : optional
-            kwargs are used to absorb the difference of parameters among the other SVD functions
-
-        Returns
-        -------
-        U : 2-D tensor, shape (matrix.shape[0], n_eigenvecs)
-            Contains the right singular vectors
-        S : 1-D tensor, shape (n_eigenvecs, )
-            Contains the singular values of `matrix`
-        V : 2-D tensor, shape (n_eigenvecs, matrix.shape[1])
-            Contains the left singular vectors
-
-        Notes
-        -----
-        This function is implemented based on Algorith 5.1 in `Finding structure with randomness:
-        Probabilistic algorithms for constructing approximate matrix decompositions`
-        - Halko et al (2009)
-        """
-        # Check that matrix is... a matrix!
-        if self.ndim(matrix) != 2:
-            raise ValueError(
-                "matrix be a matrix. matrix.ndim is %d != 2" % self.ndim(matrix)
-            )
-
-        dim_1, dim_2 = self.shape(matrix)
-        min_dim, max_dim = min(dim_1, dim_2), max(dim_1, dim_2)
-
-        if n_eigenvecs is None:
-            n_eigenvecs = max_dim
-
-        if n_eigenvecs > max_dim:
-            warnings.warn(
-                "Trying to compute SVD with n_eigenvecs={0}, which "
-                "is larger than max(matrix.shape)={1}. Setting "
-                "n_eigenvecs to {1}".format(n_eigenvecs, max_dim)
-            )
-            n_eigenvecs = max_dim
-
-        n_dims = min(n_eigenvecs + n_oversamples, max_dim)
-
-        if (
-            dim_1 > dim_2
-            and n_eigenvecs > min(min_dim, n_dims)
-            or dim_1 < dim_2
-            and n_eigenvecs < min(min_dim, n_dims)
-        ):
-            # transpose matrix to keep the reduced matrix shape minimal
-            matrix_T = self.transpose(matrix)
-            Q = self.randomized_range_finder(
-                matrix_T, n_dims=n_dims, n_iter=n_iter, random_state=random_state
-            )
-            Q_H = self.conj(self.transpose(Q))
-            matrix_reduced = self.transpose(self.dot(Q_H, matrix_T))
-            U, S, V = self.truncated_svd(matrix_reduced, n_eigenvecs=n_eigenvecs)
-            V = self.dot(V, self.transpose(Q))
-        else:
-            Q = self.randomized_range_finder(
-                matrix, n_dims=n_dims, n_iter=n_iter, random_state=random_state
-            )
-            Q_H = self.conj(self.transpose(Q))
-            matrix_reduced = self.dot(Q_H, matrix)
-            U, S, V = self.truncated_svd(matrix_reduced, n_eigenvecs=n_eigenvecs)
-            U = self.dot(Q, U)
-
-        return U, S, V
-
-    def randomized_range_finder(self, A, n_dims, n_iter=2, random_state=None):
-        """Computes an orthonormal matrix (Q) whose range approximates the range of A,  i.e., Q Q^H A ≈ A
-
-        Parameters
-        ----------
-        A : 2D-array
-        n_dims : int, dimension of the returned subspace
-        n_iter : int, number of power iterations to conduct (default = 2)
-        random_state: {None, int, np.random.RandomState}
-
-        Returns
-        -------
-        Q : 2D-array
-            of shape (A.shape[0], min(n_dims, A.shape[0], A.shape[1]))
-
-        Notes
-        -----
-        This function is implemented based on Algorith 4.4 in `Finding structure with randomness:
-        Probabilistic algorithms for constructing approximate matrix decompositions`
-        - Halko et al (2009)
-        """
-        rng = self.check_random_state(random_state)
-        dim_1, dim_2 = self.shape(A)
-        Q = self.tensor(rng.normal(size=(dim_2, n_dims)), **self.context(A))
-        Q, _ = self.qr(self.dot(A, Q))
-
-        # Perform power iterations when spectrum decays slowly
-        A_H = self.conj(self.transpose(A))
-        for i in range(n_iter):
-            Q, _ = self.qr(self.dot(A_H, Q))
-            Q, _ = self.qr(self.dot(A, Q))
-
-        return Q
-
     index = Index()
-
-    @property
-    def SVD_FUNS(self):
-        return {
-            "numpy_svd": self.partial_svd,
-            "truncated_svd": self.truncated_svd,
-            "symeig_svd": self.symeig_svd,
-            "randomized_svd": self.randomized_svd,
-        }
 
     @staticmethod
     def index_update(tensor, indices, values):
