+++ tensorly/contrib/sparse/backend/__init__.py
@@ -110,7 +110,6 @@ kron = dispatch_sparse(backend.kron)
 kr = dispatch_sparse(backend.kr)
 solve = dispatch_sparse(backend.solve)
 qr = dispatch_sparse(backend.qr)
-partial_svd = dispatch_sparse(backend.partial_svd)
 unfold = dispatch_sparse(base.unfold)
 fold = dispatch_sparse(base.fold)
 tensor_to_vec = dispatch_sparse(base.tensor_to_vec)
